+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
 ## Unreleased
+- Upgraded Otel dependencies to 1.23.0 / 0.44b0
+
+- Changed the return type of start_tracing from sdk.TracerProvider to api.TracerProvider
+- Profiling: fix the reported frame count when exporting stacktraces.
 
 ## 1.17.0 - 2024-01-22
 - Dependencies bump, removed Jaeger